NOJIRA split DomainExecutorTest into three separate tests
