+++ src/test/resources/org.junit.jupiter.api.extension.Extension
@@ -1,0 +1,1 @@
+uk.gov.justice.digital.test.SparkSessionExtension
